Pull-out Live to separate file
