--- src/main/scala/zio/rocksdb/RocksDB.scala
@@ -1,141 +1,0 @@
-package zio.rocksdb
-
-import java.{ util => ju }
-import org.{ rocksdb => jrocks }
-import zio._
-import zio.stream._
-
-import scala.jdk.CollectionConverters._
-
-object RocksDB {
-  trait Service {
-    def delete(key: Array[Byte]): Task[Unit]
-
-    def delete(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte]): Task[Unit]
-
-    def get(key: Array[Byte]): Task[Option[Array[Byte]]]
-
-    def get(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte]): Task[Option[Array[Byte]]]
-
-    def multiGetAsList(keys: List[Array[Byte]]): Task[List[Option[Array[Byte]]]]
-
-    def multiGetAsList(keysAndHandles: Map[Array[Byte], jrocks.ColumnFamilyHandle]): Task[List[Option[Array[Byte]]]]
-
-    def newIterator: Stream[Throwable, (Array[Byte], Array[Byte])]
-
-    def newIterator(cfHandle: jrocks.ColumnFamilyHandle): Stream[Throwable, (Array[Byte], Array[Byte])]
-
-    def newIterators(
-      cfHandles: List[jrocks.ColumnFamilyHandle]
-    ): Stream[Throwable, (jrocks.ColumnFamilyHandle, Stream[Throwable, (Array[Byte], Array[Byte])])]
-
-    def put(key: Array[Byte], value: Array[Byte]): Task[Unit]
-
-    def put(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte], value: Array[Byte]): Task[Unit]
-  }
-
-  def live(
-    options: jrocks.DBOptions,
-    path: String,
-    cfDescriptors: List[jrocks.ColumnFamilyDescriptor]
-  ): ZLayer.NoDeps[Throwable, Has[(RocksDB.Service, List[jrocks.ColumnFamilyHandle])]] =
-    ZLayer.fromManaged(Live.open(options, path, cfDescriptors))
-
-  def live(path: String): ZLayer.NoDeps[Throwable, RocksDB] =
-    ZLayer.fromManaged(Live.open(path))
-
-  def live(options: jrocks.Options, path: String): ZLayer.NoDeps[Throwable, RocksDB] =
-    ZLayer.fromManaged(Live.open(options, path))
-
-  final class Live private (db: jrocks.RocksDB) extends Service {
-    def delete(key: Array[Byte]): Task[Unit] =
-      Task(db.delete(key))
-
-    def delete(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte]): Task[Unit] =
-      Task(db.delete(cfHandle, key))
-
-    def get(key: Array[Byte]): Task[Option[Array[Byte]]] =
-      Task(Option(db.get(key)))
-
-    def get(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte]): Task[Option[Array[Byte]]] =
-      Task(Option(db.get(cfHandle, key)))
-
-    def multiGetAsList(keys: List[Array[Byte]]): Task[List[Option[Array[Byte]]]] =
-      Task(db.multiGetAsList(keys.asJava).asScala.toList.map(Option(_)))
-
-    def multiGetAsList(
-      keysAndHandles: Map[Array[Byte], jrocks.ColumnFamilyHandle]
-    ): Task[List[Option[Array[Byte]]]] =
-      Task {
-        val (keys, handles) = keysAndHandles.toList.unzip
-        db.multiGetAsList(handles.asJava, keys.asJava).asScala.toList.map(Option(_))
-      }
-
-    private def drainIterator(it: jrocks.RocksIterator): Stream[Throwable, (Array[Byte], Array[Byte])] =
-      ZStream.fromEffect(Task(it.seekToFirst())).drain ++
-        ZStream.fromEffect(Task(it.isValid())).flatMap { valid =>
-          if (!valid) ZStream.empty
-          else
-            ZStream.fromEffect(Task(it.key() -> it.value())) ++ ZStream.repeatEffectOption {
-              Task {
-                it.next()
-
-                if (!it.isValid()) ZIO.fail(None)
-                else UIO(it.key() -> it.value())
-              }.mapError(Some(_)).flatten
-            }
-        }
-
-    def newIterator: Stream[Throwable, (Array[Byte], Array[Byte])] =
-      ZStream
-        .bracket(Task(db.newIterator()))(it => UIO(it.close()))
-        .flatMap(drainIterator)
-
-    def newIterator(cfHandle: jrocks.ColumnFamilyHandle): Stream[Throwable, (Array[Byte], Array[Byte])] =
-      ZStream
-        .bracket(Task(db.newIterator(cfHandle)))(it => UIO(it.close()))
-        .flatMap(drainIterator)
-
-    def newIterators(
-      cfHandles: List[jrocks.ColumnFamilyHandle]
-    ): Stream[Throwable, (jrocks.ColumnFamilyHandle, Stream[Throwable, (Array[Byte], Array[Byte])])] =
-      ZStream
-        .bracket(Task(db.newIterators(cfHandles.asJava)))(its => UIO.foreach(its.asScala)(it => UIO(it.close())))
-        .flatMap { its =>
-          ZStream.fromIterable {
-            cfHandles.zip(its.asScala.toList.map(drainIterator))
-          }
-        }
-
-    def put(key: Array[Byte], value: Array[Byte]): Task[Unit] =
-      Task(db.put(key, value))
-
-    def put(cfHandle: jrocks.ColumnFamilyHandle, key: Array[Byte], value: Array[Byte]): Task[Unit] =
-      Task(db.put(cfHandle, key, value))
-  }
-
-  object Live {
-    def listColumnFamilies(options: jrocks.Options, path: String): Task[List[Array[Byte]]] =
-      Task(jrocks.RocksDB.listColumnFamilies(options, path).asScala.toList)
-
-    def open(
-      options: jrocks.DBOptions,
-      path: String,
-      cfDescriptors: List[jrocks.ColumnFamilyDescriptor]
-    ): Managed[Throwable, (RocksDB.Service, List[jrocks.ColumnFamilyHandle])] = {
-      val handles = new ju.ArrayList[jrocks.ColumnFamilyHandle](cfDescriptors.size)
-      val db      = Task(jrocks.RocksDB.open(options, path, cfDescriptors.asJava, handles))
-
-      make(db).map(_ -> handles.asScala.toList)
-    }
-
-    def open(path: String): Managed[Throwable, RocksDB.Service] =
-      make(Task(jrocks.RocksDB.open(path)))
-
-    def open(options: jrocks.Options, path: String): Managed[Throwable, RocksDB.Service] =
-      make(Task(jrocks.RocksDB.open(options, path)))
-
-    private def make(db: Task[jrocks.RocksDB]): Managed[Throwable, RocksDB.Service] =
-      db.toManaged(db => Task(db.closeE()).orDie).map(new Live(_))
-  }
-}
